Some more z series warning cleanup
git-svn-id: https://svn.code.sf.net/p/oorexx/code-0/main/trunk@4758 0b6cbdbe-3aab-466e-b73a-abd511dda0a2

+++ common/platform/unix/SysFile.cpp
@@ -831,7 +831,7 @@ bool SysFile::setPosition(int64_t location, int64_t &position)
     if (location >= (filePointer - bufferedInput) && location < filePointer)
     {
         // just shift the buffer position;
-        bufferPosition = location - (filePointer - (int64_t)bufferedInput);
+        bufferPosition = (size_t)(location - (filePointer - (int64_t)bufferedInput));
         // just return the same value
         position = location;
     }
